requirements.txt
GitPython==2.1.7
pytest==3.6.1